Now updates refs instead of merging pull requests
I didn't like how all the child repos ended up with a bunch of merge
commits that the parent repo didn't have. I wanted them to look exactly
the same. So now we still create a PR on the child repo, but that's
solely to get the hash from the parent to exist in the child repo, so
that we can set the ref to that hash